+++ crates/provider/src/provider/mod.rs
@@ -19,5 +19,4 @@ pub use wallet::WalletProvider;
 mod with_block;
 pub use with_block::{ParamsWithBlock, RpcWithBlock};
 
-mod caller;
-pub use caller::Caller;
+pub use eth_call::Caller;
